Remove rewriting for universe set (#9979)
The rewrites had led to complications with the cardinality solver and moreover weren't sound if the universe set was e.g. intersected with a term that was not a subset of elements from set variables.

This furthermore corrects an issue cvc5/cvc5-projects#644, which was triggered by the unsoundness in these rewrites.

Fixes cvc5/cvc5-projects#644.