Rename requirePhase to preferPhase. (#9982)
The name of function SatSolver::requirePhase() (and friends) is misleading. It implies that the configured phase is always enforced, which is not the case: it is only enforced when the SAT solver decides to make a decision on the given variable on its own, or if the decision was derived by a DecisionEngine. Decisions derived by a theory ignore this configure phase (theories require that the phase they require overrides the phase configured via requirePhase()). Thus it is now renamed to preferPhase().