Change syntax for unit tuples (#10012)
Also fixes previous documentation for tuples.

This is in preparation for the alf proof backend. The current use of Tuple to denote both unit and non-unit tuples is highly ambiguous. This makes UnitTuple the new smt2 syntax for unit tuples.

The PR currently adds support but does not delete support for the old syntax (both can be used), although I am open to changing the syntax to the new version permanently.

This PR also adds several missing cases of smtKindString.