Allow free assumptions from input when merging subproofs (#9886)
This enables better proof compression at post-processing time.

The idea is to allow any proof contain free assumptions that are a subset of the input as a source for merging. This is in contrast to our previous policy that only used proofs with no free assumptions as sources for merging.

One consequence is that timeout and unsat cores can be smaller when --proof-granularity=theory-rewrite. This is because proofs involving substitutions are now made fine grained, e.g. we can now replace:
(and_elim i (and_intro P1 .... Pn)) ----> Pi
where P1 ... Pn are proofs of the globabl substitution of equality. Previously any proof involving substitutions would depend on all assertions that contributed to a substitution.

In the regressions, this change leads to a 10% reduction in overall proof size.