[proof-new] Add test cases to exercise bitvector proof reconstruction rules. (#10072)
Remove bvurem and bvudiv RARE rules that handle negative cases - they are never invoked. More specifically, the corresponding rules UremPow2 and UdivPow2 require the divisor to be non-negative. The negative cases are removed, and the cases for dividing by 1 are handled by bv-{udiv,urem}-one.
Add some test cases that are supposed to be filled by exisitng rules, but either a roundabout method is used or the hole doesn't get filled.
Fixed a bug in bv-usubo-eliminate
Fixed a bug in bv-extract-mult-leading-bit
There remain 3 cases of bv reconstruction rules unable to fill holes generated by the rewriter.