Make sorts introduced by ho elim and fmf-fun have a consistent name (#10018)
In preparation for the alf backend.

This ensures sorts introduced by ho elim and fmf-fun can be defined in the cvc5 signature.

In particular, previously we introduced fresh sorts as necessary in these preprocessing passes. These sorts would not be accounted for in the ALF proof, since we do not traverse the proof to find sorts not appearing in the input.

Now, we define a family of sorts where the sorts introduced in the preprocessing passes are instances of these sorts. These families of sorts will have a proper definition in the ALF signature.