Partition Generator Updates Part 1 (#10099)
This PR updates a few naming schemes but mainly focuses on adding in the ability to emit pending partitions. Emitting pending partitions is important for any scatter strategy because while the user may request 64 partitions, it is possible that through blocking the search, only 40 (or some other number less than 64) partitions are created before the solver arrives at some solution. In this case, we need to dump all of the partitions that were created.

Note that d_emittedAnyPartitions is unused for now, but it will be used in future PRs.