Make CoCoA use cvc5's libgmp path
Prior this PR, CoCoA tries to discover libgmp on its own.
This will usually fail on systems like Mac M1/2, where libgmp via
Homebrew installation is usually located at a non-standard location.
This PR passes an explicit libgmp path that cvc5 already computed
to CoCoA, which fixes the issue.

We need to also manually patch CoCoA because CoCoA is buggy:
it made an assumption that libgmp is statically-linked,
so when cvc5 is compiled in the dynamically linked mode
(i.e. without `--static`), CoCoA will fail to compile.
The patch fixes this issue.

Signed-off-by: Sorawee Porncharoenwase <sorawee.pwase@gmail.com>