FieldObj class & CoCoA utilities (#10158)
The FieldObj class is meant to be inherited by a subclass which is specific to some finite field. It includes stores the field size, common field constants, and helper functions that depend on the field.

This is helpful b/c with the new solver, we end up with a few different classes that are all field-specific.

This PR also adds a few other CoCoA-related utilities that it will be helpful to have centralized.