More fixes for HO + MBQI (#10206)
Includes:

Simplifies HO term indexing by always asserting unit clauses corresponding to the necessary purification eagerly. The previous implementation could mark that we are in conflict while sending a lemma that already is true in the current context, leading to wrong "sat" answers.
Catching more cases of when a term fails to evaluate in a model.
Fix MBQI for model values that include uninterpreted constants embedded in function array constants.
Note this also adds a useful rewrite for lambdas that helps solve some of the tested problems faster.