prop: cadical: Fix check_found_model_handling() when rejecting without reason. (#10242)
When checking the SAT model, it can happen that we introduce new
variables via getNextDecisionRequest() or theoryCheck(). If so, we force
CaDiCaL to give us a model for these new variables before we actually
check the model for consistency. We rely on CaDiCaL here to kick off the
whole assignment notification machinery. At the moment (for the
currently configured CaDiCaL 1.7.4 up to version 1.9.3), CaDiCaL expects
us to provide a reason clause when we reject the model, which we don't
have for these cases. For now, until this assertion is weakened in
CaDiCaL, for these cases we pacify it with a tautology instead.