Miniscope top-level AND in assertions pipeline (#10208)
This makes it so that top-level AND assertions are immediately miniscoped upon insertion in the assertions pipeline.

The motivation for this is to minimize proof dependencies. In particular, currently for an input:
(assert (and A B C))
(assert (not B))

Our proof is roughly:

Rewriting (and A B C) to (and A' B' C'), B to B'.
And-elim to extract B' from (and A' B' C')
Conclude false
The key point is that the rewriting A -> A' and C -> C' is irrelevant.

This furthermore eliminates any use of "conjoin" in preprocessing for similar reasons.

With this change, our proofs become roughly 6% smaller (measured on regressions). Our overall proof overhead decreases 4.4x -> 4.1x in my latest experiments.