Simplify interface for skolem functions (#10179)
This is another step towards an external interface for skolem functions. It introduces an internal method for getting the type of a skolem function so that the user does not need to supply one.

Some skolem functions are internal and are not intended to be shared. For these, a new variant mkSkolemFunctionTyped is introduced, where a type is manually provided.

With this change, the syntax for (most) skolems do not need to carry their types.