Refactor proofs in the proposition layer (#10083)
This PR simplifies the relationships between several components in the prop layer that are used for proofs.

In particular:

SatProofManager is now specific to Minisat
Several components of ProofCnfStream are distributed to the proper place. ProofCnfStream is now fully encapsulated in PropPfManager.
PropPfManager now is the only interface to proofs from PropEngine
TheoryProxy does not make direct calls to ProofCnfStream.
All things involving optimized proof managers are moved to SatProofManager
Further PRs will move the minisat-specific files to src/prop/minisat.