Fix Python package licensing (#10348)
This PR updates the `License` field of the Python wheel metadata to
reflect that the package includes artifacts under three different
licenses: BSD-3-Clause, LGPL-3.0-or-later, and MIT. It also includes a
COPYING file and associated license files in the `dist-info` directory
of the Python wheel.

It fixes #10327. It depends on #10339.