Separate external and internal portions of skolem API (#10336)
Review #10357 first.

This is work towards exporting skolem ids in the API.

This PR distinguishes two kinds of skolems:
(1) ones that we will export in the API,
(2) "internal" ones that we will not support in the API.

This PR makes it mandatory for skolems of type (1) to have typing rules formalized. Conversely, rules for (2) are not formalized. This PR updates the interface of SkolemManager accordingly.

Note that SHARED_SELECTOR is converted in a separate PR #10357.