Fix version detection code for CLN >= 1.3.7 (#10372)
The format of CL_VERSION in CLN's version.h changed in v1.3.7. Before v1.3.7, the version was available in a single line. Now, the major, minor, and patchlevel version numbers are specified in three separate lines. This PR fixes FindCLN.cmake so that the version number is parsed properly in all versions of the library.

Fixes #10370.