Do not mark clauses as removable. (#10396)
Marking clauses as removable in MiniSat was introduced prior to the introduction of the equality engine (a long time ago). As it is now, marking clauses as removable has a huge negative impact on performance, especially on strings. The question of when to mark which clauses as removable is currently being investigated from scratch.