Use setuptools to build Python bindings (#10455)
This PR replaces scikit-build with setuptools to build the Python
bindings. Our setup with scikit-build didn't work on Windows, and used
the FindPythonInterp and FindPythonLibs modules that were removed in
CMake 3.27 (see policy
[CMP0148](https://cmake.org/cmake/help/v3.27/policy/CMP0148.html)). This
PR also uses pip to install the Python bindings using the Wheel format
instead of using the deprecated call to `setup.py install
--single-version-externally-managed`, which used the obsolete .egg
format.

A follow-up PR will leverage this new setup to package the Python
bindings for all the supported platforms and Python implementation
versions using [cibuildwheel](https://cibuildwheel.readthedocs.io/). In
this PR, only a minimal number of changes to the current packaging
scripts have been applied to keep the packaging process working in the
meantime.