Refactor sygus expanded definition form (#10399)
This improves how "expanded definition form" is computed for our uses of sygus.

This marks the form of sygus operators after expanding definitions, so that we do symmetry breaking modulo rewriting wrt operators after expanding user definitions.

Previously it was only computed by the sygus solver at the beginning of check-synth. Now it is done after grammar normalization in the internal sygus solver (which makes us more robust to when sygus operators involving definitions change), and in the SygusTermEnumerator, meaning that any user of the fast enumerator does not have to worry about this detail.

It furthermore updates the find-synth utilities to use SygusTermEnumerator so that this detail is properly handled.