Bump CMake version to 3.16 (#10468)
This PR raises the minimum required version of CMake to 3.16. This aligns with the CMake version used in our CI environment on Ubuntu 20.04. Additionally, to improve the detection of CMake version compliance issues across different configurations, this PR also adjusts all Ubuntu runners to version 20.04. This is needed since only the portions of the CMake code that are executed are tested for version compliance.