Generalize decision engine to accept local assertions (#10453)
This refactors the decision engine to accept "local assertions", which are formulas whose scope is SAT-context dependent.

The functionality for such lemmas was already available via "skolem definitions" which dynamically become active. This PR makes it so that this functionality was also be used for theory lemmas.

This functionality is not yet leveraged, thus there are no behavior changes in this PR.

It also makes some minor simplifications where we assumed that "removable" lemmas were to be ignored by the decision heuristic.