Do not invoke rewriter on terms involving abstract sorts in DSL proof reconstruction (#10517)
This adopts the policy that terms involving abstract sorts should not be passed to the rewriter. We consider terms with any abstract subterms to rewrite to themselves in the DSL proof reconstruction algorithm.

Pros:

This makes it so that we do not need to make the rewriter robust to rewriting with abstract subterms. This is particularly important for BV which assumes that all children of BV operators have concrete bitwidth. This change makes it so that the BV rewriter does not require further changes from proof-new. The same applies to the FP rewriter.
Cons:

The DSL reconstruction algorithm is not able to invoke the rewriter as an oracle in some contexts. This means that certain optimizations are not applicable, especially when doing proof reconstruction for BV.
In my alf development branch, this limitation leads to only a difference of 2 unproven BV rewrites on our regressions (99 -> 101).

This also includes a minor change from proof-new to ensure terms are properly converted upon entry to the DSL rewrite reconstruction algorithm.

This fixes make regress-dsl-proof and should fix the nightly failures.