Do not use mixed arithmetic in type rules for transcendentals (#10536)
Type rules for transcendentals were originally based on the operators for linear arithmetic which permit mixed arithmetic. There is no benefit in allowing this.

Work towards addressing #10534. The legal variant of the benchmark from that issue is now:

(declare-const a Int)                                            
(assert (> (tan (sin (sin (to_real a)))) a))
(check-sat)
This still reproduces the incorrect "unsat" after this PR, which will be addressed separately.