Make skolems for all extended transcendental functions have functional type (#10555)
This ensures e.g. arccos is a function on intervals where it is undefined. This corrects invalid models on the 2nd and 3rd benchmarks #10542. The solution is to generalize how SQRT is handled to all inverse transcendental functions.

It also adds the regression from #10541, which was a prior issue in cvc4. Fixes #10541.

Also changes the rewriter for exp slightly, which was bogus due to type issues, also adds (exp 0.0) --> 1.0, which fixes a warning from #10543.