Implement printing for skolems (#10563)
This PR adds an option --print-skolem-definitions (disabled by default) to print skolems based on their formal definitions.

This will eventually be used when exporting terms via the text interface.

It also makes a minor modification to QUANTIFIERS_SKOLEMIZE to take the variable instead of the index, as this corrects a special case in the ALF proof signature. FYI @HanielB

Furthermore skolem names are now lowercase (prefixed by @) for consistency with other operators.