ci: Set macOS version with MACOSX_DEPLOYMENT_TARGET (#10525)
This PR uses the MACOSX_DEPLOYMENT_TARGET variable in CI to set the
minimum required macOS version for the cvc5 libraries and binaries,
which is 10.13 for Intel and 11.0 for Apple Silicon. With this change,
we can use newer versions of the macOS runners to build cvc5 while
keeping compatibility with older macOS versions. To ensure the GMP
library also respects the target macOS version, we force its compilation
from sources.