Fixes for HO model building (#10526)
This fixes several issues with HO model construction which arose while working on a new technique for MBQI+HO.

In particular:
(1) There was a discrepancy between which function terms were "assignable". One place in the code said any non-lambda was assignable, and one place said that only function variables were assignable. This is now syncronized by isAssignableUf(...).
(2) We did not mark lambdas as the "UF model" in higher-order in cases where a lambda was in an equivalence class of function type. This PR introduces assignRepresentative(...) to ensure we account for this.
(3) We did not recursively normalize lambda values based on the model values for symbols they rely on. This ensures that we return fully evaluated values for function terms in HO logics. A regression is added for this issue.