Update ALF signatures to parameterized nil terminators (#10566)
This updates the alfc version to one supporting declare-parameterized-const and deprecating associative operators with unspecified nil terminators. It also deprecates alf.null.

This change is necessary for the ALF+RARE integration moving forward.

Note the style of side conditions in the cvc5+ALF signature is updating to always use $, this adopts the new style and does some minor cleanup to existing side conditions.