Move DSL rewrite rule reconstruction to separate module (#10450)
This splits DSL rewrite rule reconstruction for the main proof postprocessor for clarity and flexibility.

Trusted proof steps are now collected and then reconstructed as a second step.

We can additionally use this class to fill in holes introduced after post-processing (e.g. the elim subtypes pass). This will be done on a separate PR.