Updates to the BV theory in ALF (#10616)
This is in preparation for RARE+ALF.

This updates the type rules for certain BV operators to take into account evaluation on indices.

This is essentially making our ALF signature reflect the change in #10598.

This ensures that the RARE rules involving e.g. extract with symbolic indices can be used as-is.

It also improves the consistency of our utilities for nil terminators by making "@bv_empty" a skolem to refer to the bitvector constant 0.