Use standard order for equality rewriting in BV (#10641)
This updates BV's rewriter to use the standard order for rewriting equalities used by all other theories.

This avoids non-termination in the rewriter exhibited in RARE proof reconstruction after 4a77a14.

This change is already on proof development branches. This should fix the nightlies.

This may impact performance, but regardless is required for correcting the issue.