[proofs] Fix handling of cycles in LazyProofChain (#10683)
Previously when finding a cycle in a chain at LazyProofChain, where the proof a given F would depend on G whose proof would have F as a free assumption, we were preventing the creation of a cyclic proof for G, as expected, but the partial proof of F in terms of G was still being created. This commit fixes this behavior such that G's bogus proof will be discarded, as before, but also F's.

The other required fix is to avoid creating resolution links to literals/clauses that we have a justification for in the CNF proof. This will avoid introducing spurious resolution steps but also introducing potential cycles in the lazy proof chain.

This commit also changes one trace in the Minisat proof manager to print proofs with conclusions.

This fixes cvc5/cvc5-projects#711