Install Python bindings in interpreter default location (#10693)
Users expect the Python bindings to be installed in the default location designated by the Python interpreter used to build them (see discussion #10255), and not the default system path set by CMake that may be not associated with any Python version. This changes the installation step to meet their expectation. Furthermore, it simplifies the installation code and ensures consistent behavior across Python environments, regardless of whether they are virtual or not.