Flatten quantified formulas introduced for negative REGEXP unfolding (#10684)
This modifies quantified formulas introduced for negative REGEXP unfolding so that they do not contain nested Boolean structure, i.e. (=> (and A B) C) is converted to (or (not A) (not B) C).

Note these rules have yet to be formalized so there is not an ALF signature for these yet.

This avoids the need for rewrite steps in the bodies of quantified formulas in final proofs, which moreover RARE fails to find in some cases.