Simplify policy for Boolean term skolems (#10711)
This PR cleans up our policy for purification skolems.

In particular, we have been using purification for two purposes:
(1) Boolean purification skolems indicate that the variables should be treated as theory atoms for theory combination. This meant that Theory::theoryOf checked whether a variable has been marked as a purification skolem, thus making it belong to THEORY_UF.
(2) purification skolems are used as a way of marking the definition of internally introduced variables, making our proof system trivially justify the replacement of terms by their purification skolem.

These two use cases clashed. In particular, if one introduced a purification skolem for any term in a logic that did not include UF, a logic exception would occur. Moreover this meant that using a purification skolem in a subsolver would be taken out of context as a theory atom.

This makes Theory::theoryOf independent of whether a variable is a purification skolem, and instead tracks this set explicitly in Env.

This will enable more use cases of purification skolems to justify proofs of preprocessing.