Support `ProofRule::CONCAT_LPROP` in ALF (#10661)
This also changes a utility in the core solver which did a micro-optimization that could lead to proof failures. This optimization was to handle cases where both (x,y) and (y,x) were split on so that you would get a common Skolem. This should never happen in practice and moreover the definition of the Skolem would not be its expected form.

This PR also updates the documentation for CONCAT_LPROP to be accurate wrt the optimized form.