Add support for set.is_empty (#10743)
As discussed in the proofs meeting, RARE cannot express empty set (without referring to type parameters). The solution is to introduce a predicate to check whether or not a set is empty. This introduces this as a first class kind to our API.

Follow up PRs will use this to write conditional RARE rules to express rewrites involving empty set.