cmake: Do not search for Python development artifacts (#10757)
Specifying the Development component in CMake for Python is unnecessary.
Even if the search for the Development component fails, it doesn't
affect the successful compilation of Python bindings. However, not
finding the development artifacts might terminate the build prematurely.
This situation is currently observed when building Python bindings with
cibuildwheel in the PyPi packaging action. Therefore, this updates the
CMakeLists file to exclude the Development component during the Python
location process.