Update ProofRule::THEORY_REWRITE to take an equality (#10764)
This is required since some rewrite rules (e.g. strings arithmetic approximations) are easy to check but hard to compute.

In particular, a forthcoming rewrite will invoke the search procedure from https://homepage.divms.uiowa.edu/~ajreynol/cav19c.pdf, but the proof checker can be configured to verify that the chosen approximation is one of the allowed ones.