Fixes several issues related to DSL/THEORY rewrites (#10774)
ncludes:

RARE reconstruction for fixed point rules was broken in a fairly major way, since we could introduce variable aliasing for fixed point contexts. The fix is to canonize context terms at the same time as conditions/conclusions, as this will ensure all variables are unique.
We weren't tracking the DSL rewrite id in some cases of fixed point reconstructions
Fixes the (internal) printer for THEORY_REWRITE
Increases the step limit of RARE reconstruction from 200 -> 1000, which is important for the industrial set of interest, and appears to lead to better performance on regressions.
The first issue would manifest as open proofs when doing RARE reconstruction. The second issue would manifest as segfaults.