Handle concatenations with variables in regexp consume utility (#10791)
This makes the simple regular expression consume utility handle (non-value) concatenations within str.to_re.

This change was uncovered when attempting to write a proof checker for this utility, where the proof checker was capable of more powerful rewrites since it did not distinguish value vs. non-value.

This is in preparation for a new theory rewrite ProofRewriteRule::STR_IN_RE_CONSUME which will be based on this utility.

This change leads to faster solving on a few industrial problems of interest.

Adds several regressions which excercised corner cases of this reasoning in the proof checker.