Add infrastructure for elaboration of MACRO theory rewrites (#10803)
This adds infrastructure for generating proofs for THEORY_REWRITE, where if the rewrite is a MACRO_ rule, we run an elaboration step in the rewrite rule proof constructor (post-RARE).

If the proof of the elaboration itself involves holes, then these holes are recorded as "subgoals" during RARE reconstruction. The proof post-processor then applies a second step to fill in these proofs using further calls to RARE reconstruction.

This adds support for eliminating MACRO_BOOL_NNF_NORM in this manner. In particular, the proof generation for NNF norm will add many small steps which themselves correspond to small step rewrites, in particular the RARE rules from #10782 are targeted for this elaboration.