Refresh assertions for get-abduct and get-interpolant (#10252)
This ensures e.g. variables that can be eliminated during preprocessing don't show up in the goal/grammar for abduction and interpolation.

There are pros and cons to doing this, however it should be noted that this is a more consistent behavior since calling check-sat prior to a call to get-abduct would have the same effect.

Fixes #10788.