Simplify ENCODE_PRED_TRANSFORM (#10799)
This PR does two things:
(1) It makes a necessary change to preprocessing terms before RARE reconstruction that applies ACI_NORM recursively to all subterms. This is required for proving rewrites involving large string constants, which after "unflattening" are not in ACI normal form.
(2) It simplifies the definition of ENCODE_PRED_TRANSFORM to refer to the *small step* rewrites of this utility, not the large step.  It make the converter proof producing for the small steps using a conversion proof generator.

With this change `ENCODE_PRED_TRANSFORM` is now essentially equivalent to `REFL`. 

The new name of this rule is `ENCODE_EQ_INTRO`.