Refactor strings arithmetic entailment utility (#10768)
Includes new helper interfaces and adds interface that will be critical for proof support.

Also makes it so that the rewriter is optional, since we cannot rely on the rewriter when proofs are enabled.

We also can't rely on the search procedure in the proof checker, so we pass a flag isSimple to many methods to ensure we use only "simple" methods for provide when an approximation is sound. This is restricted only if we are attempting a proo rewrite step.

This is in preparation for two new THEORY_REWRITE rules for strings.