Specify Threads as a dependency explicitly (#10790)
Building cvc5 with CryptoMiniSat generates a dependency on the thread library. The config file cvc5Targets.cmake automatically generated by CMake didn't reflect this because the dependency was not explicitly specified. It was relying on passing the -pthread compiler and linker flag. This PR explicitly adds Threads as a dependency.