Eliminate witness from BV invertibility instantiations (#10848)
This makes an explicit pass to eliminate WITNESS terms from instantiations in the instantiation strategy from Niemetz et al CAV 2018.

The motivation is two-fold:
(1) It is work towards eliminating an incorrect dependency on proofs from SkolemManager. Proofs are local to solving instances, while SkolemManager has a lifetime that is equivalent to its containing NodeManager.
(2) It will make proofs easier to track from the BV instantiator later, as they now can be given on individual lemmas instead of relying on a central utility.

A followup PR will eliminate support for WITNESS in skolem manager and the term formula removal pass.