Eliminate handling of WITNESS in assertions (#10861)
This eliminates the dependency on proofs from the skolem manager, which could lead to issues where skolems are created in the same skolem manager in different solvers (with/without proofs).

This dependency was to support a deprecated feature where proofs could be attached to witness terms. This functionality was never used and moreover complicated some of our proof rules and the term formula removal preprocessing pass. This can be simplified now.

This is work towards stabilizing the ALF signature as well as supporting more advanced usages of skolems in distributed/parallel cvc5 instances.