Fix issue in beta reduction and add support in ALF (#10879)
This adds support for ProofRewriteRule::BETA_REDUCE in ALF.

When formalizing this rule, three issues arose:

(1) We do not properly model variables from define-fun, thus beta-reduce was not properly checkable in the ALF signature. This PR changes our smt2 parser to use canonical variables, similar to how the parser for quantified formulas was updated here: #10556.

(2) When fixing this discrepancy, a bug was uncovered in the UF rewriter. In particular, the cvc5 rewriter does not rewrite terms that appear in operators (this occurs rarely, only when there are beta-redexes or we have an HO logic). This meant that it was theoretically possible for a lambda with variable shadowing to appear in an APPLY_UF term, thus leading to a bogus beta reduction when rewritten. This issue could only appear in raw API usages where a user could manually construct a lambda with shadowing and put it in an APPLY_UF.

This PR fixes this issue by ensuring lambdas are rewritten prior to performing beta reduction. At the moment, this requires a backwards reference to the rewriter from the UF rewriter (this is not an ideal practice, but a similar pattern is used in some of our other theory rewriters).

(3) Our shadow elimination was not considering the case where a variable occurs multiple times in the same quantifier prefix. This is also fixed in this PR.

A more permanent solution will be to rewrite inside of operators by default in the core rewriter, although this is a major change.