Switch AttrHash<V> from std::unordered_map to bespoke two-level map (#10648)
Performance on my test case (described in discussion #10578) is dominated by NodeManager::reclaimZombies() and in particular, deleting all attributes matching a particular NodeValue. A straightforward way to improve this is to have a two-level structure where the top is indexed by NodeValue*, giving O(1) deletion. Using a flat map for the second level (indexed by uint64_t) fixes the increased memory usage of the first step and gives some additional performance. The net result is more than a factor of two speedup