Share theory lemmas and SAT clauses containing skolems with plugins (#10901)
This updates the policy to share theory lemmas and SAT clauses
containing skolems with plugins by default.

The old behavior can be obtained by `--no-plugin-share-skolems`.

This also moves the implementation of getSharableFormula to Env since it
depends on options now, and consolidates plugin options in the
`base_options.toml` file for consistency.