ff: bugfix: CoCoA tracer
We give CoCoA polynomials P = {p1, ..., pN} and ask for a GB.
We record poly ops that it does to get the GB.
If the GB has 1, we trace the ops to find a subset Q of P that gives 1.
If we miss a poly op, then the trace fails.
The bug is a missing poly op: normalizing the elements of P to be monic.
However, it's not clear how to trace CoCoA's normalizations.
So now, we normalize ourselves, before calling GB.

fixes #10937