Fix ProofRule::STRING_DECOMPOSE and add to ALF signature. (#10894)
This is an infrequently used rule for handling string disequalities.

It was not covered in the nightlies (since proof elaboration in the
strings solver failed), and moreover its internal proof checker was
bogus.

Note that the premise of the lemma was expected to be GEQ, not GT. This
weakens the premise of the lemma sent in the core solver, which matches
the proof rule (and thus is sound).

This fixes these issues and adds its definition to ALF.

---------

Co-authored-by: Abdalrhman Mohamed <abdoo8080@outlook.com>