Add macro proof rewrite rule MACRO_SUBSTR_STRIP_SYM_LENGTH (#10844)
This rule is required to discover the proper split of a string for which
we strip portions of it based on length entailment tests.

It works in tandem with three new RARE rules added in
https://github.com/cvc5/cvc5/pull/10832 (str-substr-len-include,
str-substr-len-include-pre, str-substr-len-skip). These rules were
written in a way that does not require the RARE reconstruction algorithm
to find the appropriate split, as this does not scale. Instead, the
elaboration of this macro simplifies it by explicitly by making the
split explicit so it will be easy for the RARE reconstruction to find.